+++ src/dcos_e2e/cluster.py
@@ -79,6 +79,7 @@ class Cluster(ContextDecorator):
             )
             raise ValueError(message)
 
+        self._default_ssh_user = cluster_backend.default_ssh_user
         self._destroy_on_error = destroy_on_error
         self._destroy_on_success = destroy_on_success
         self._log_output_live = log_output_live
@@ -108,6 +109,7 @@ class Cluster(ContextDecorator):
         """
         Wait until DC/OS has started and all nodes have joined the cluster.
         """
+
         diagnostics_args = [
             '/opt/mesosphere/bin/dcos-diagnostics',
             '--diag',
@@ -117,8 +119,10 @@ class Cluster(ContextDecorator):
         ]
 
         for node in self.masters:
-            node.run_as_root(
+            node.run(
                 args=diagnostics_args,
+                # Keep in mind this must be run as privileged user.
+                user=self.default_ssh_user,
                 log_output_live=self._log_output_live,
                 env={
                     'LC_ALL': 'en_US.UTF-8',
@@ -174,6 +178,13 @@ class Cluster(ContextDecorator):
         """
         return self._cluster.public_agents
 
+    @property
+    def default_ssh_user(self) -> str:
+        """
+        Return the default SSH user for accessing a node.
+        """
+        return self._default_ssh_user
+
     def run_integration_tests(
         self,
         pytest_command: List[str],
@@ -222,8 +233,9 @@ class Cluster(ContextDecorator):
         # Tests are run on a random master node.
         test_host = next(iter(self.masters))
 
-        return test_host.run_as_root(
+        return test_host.run(
             args=args,
+            user=self.default_ssh_user,
             log_output_live=self._log_output_live,
             env=environment_variables,
         )